requirements.txt
beautifulsoup4==4.9.3
geopy==2.0.0
mysql-connector-python==8.0.21
nltk==3.5
numpy==1.19.2
pymorphy2==0.9.1
pymystem3==0.2.0
pyTelegramBotAPI==3.7.2
python-dateutil=2.8.1
pyyaml==5.3.1
selenium==3.141.0
tqdm==4.51.0
whenareyou==0.3.2
